turing: fix freemapentry function
This function recursively called itself in a loop. There is no need to
that. Keep in mind though, that we will never call freemapentry() on an
entry where entry->next != NULL. This is the case because freemapentry()
is currently only called when setval fails and this is only the case if
a new entry wasn't added because an entry with the same key already
existed.